refactor(backend): resolve dependency issues
Motivation
----------
This is a by-product of #2446.

Implementing a reasonable error handling e.g. with logging leads to
a circular dependency. This PR resolves a number of things:

* circular dependency config.ts -> logger.ts -> config.ts
* non-isolated tests (order of execution matters)
* mutation of global state (CONFIG + process.env)
* mocking of code that is actually under test
* brevo client is passed to `context`

How to test
-----------
1. `cd backend`
2. `npm run test:unit`
3. Tests pass

**squashed commits**
- refactor(backend): resolve dependency issues
