+++ backend/src/config/config.ts
@@ -5,8 +5,6 @@ import { config } from 'dotenv'
 // eslint-disable-next-line import/named
 import { v4 as uuidv4 } from 'uuid'
 
-import { printConfigError } from './printConfigError'
-
 // Load env file
 config({
   path: path.resolve(__dirname, '../../.env'),
@@ -19,6 +17,7 @@ const toNumber = (env: string | undefined): number | undefined => {
 
 // Config
 const {
+  NODE_ENV,
   BREVO_KEY,
   BREVO_ADMIN_NAME,
   BREVO_ADMIN_EMAIL,
@@ -75,6 +74,7 @@ export const CONFIG = {
   ...BREVO,
   ...BBB,
   ...FRONTEND,
+  NODE_ENV,
   WELCOME_TABLE_MEETING_ID,
   WELCOME_TABLE_NAME,
   JWKS_URI,
@@ -82,49 +82,3 @@ export const CONFIG = {
   SENTRY_ENVIRONMENT,
   WEBHOOK_SECRET,
 }
-
-// Config Checks
-export const CONFIG_CHECKS = {
-  CONFIG_CHECK_BREVO_SEND_CONTACT: (
-    config: typeof CONFIG,
-  ): config is typeof CONFIG & {
-    BREVO_KEY: string
-    BREVO_CONTACT_TEMPLATE_ADMIN: number
-    BREVO_CONTACT_TEMPLATE_USER: number
-    BREVO_ADMIN_NAME: string
-    BREVO_ADMIN_EMAIL: string
-  } =>
-    typeof config.BREVO_KEY === 'string' &&
-    typeof config.BREVO_CONTACT_TEMPLATE_ADMIN === 'number' &&
-    typeof config.BREVO_CONTACT_TEMPLATE_USER === 'number' &&
-    typeof config.BREVO_ADMIN_NAME === 'string' &&
-    typeof config.BREVO_ADMIN_EMAIL === 'string',
-  CONFIG_CHECK_BREVO_NEWSLETTER: (
-    config: typeof CONFIG,
-  ): config is typeof CONFIG & {
-    BREVO_KEY: string
-    BREVO_NEWSLETTER_LIST: number
-    BREVO_NEWSLETTER_TEMPLATE_OPTIN: number
-    BREVO_ADMIN_NAME: string
-    BREVO_ADMIN_EMAIL: string
-  } =>
-    typeof config.BREVO_KEY === 'string' &&
-    typeof config.BREVO_NEWSLETTER_LIST === 'number' &&
-    typeof config.BREVO_NEWSLETTER_TEMPLATE_OPTIN === 'number' &&
-    typeof config.BREVO_ADMIN_NAME === 'string' &&
-    typeof config.BREVO_ADMIN_EMAIL === 'string',
-}
-
-const validateConfig = () => {
-  if (!CONFIG_CHECKS.CONFIG_CHECK_BREVO_SEND_CONTACT(CONFIG)) {
-    printConfigError(
-      'BREVO_SEND_CONTACT functionality is disabled - some BREVO configs are missing',
-    )
-  }
-
-  if (!CONFIG_CHECKS.CONFIG_CHECK_BREVO_NEWSLETTER(CONFIG)) {
-    printConfigError('BREVO_NEWSLETTER functionality is disabled - some BREVO configs are missing')
-  }
-}
-
-validateConfig()
