+++ backend/src/graphql/resolvers/ContactFormResolver.spec.ts
@@ -1,41 +1,43 @@
 import { ApolloServer } from '@apollo/server'
 
-import { sendContactEmails } from '#api/Brevo'
 import { prisma } from '#src/prisma'
 import { createTestServer } from '#src/server/server'
+import { deleteAll } from '#test/helpers'
+import { mockContextValue } from '#test/mockContextValue'
 
 import type { Context } from '#src/context'
 
 let testServer: ApolloServer<Context>
-
-// eslint-disable-next-line jest/no-untyped-mock-factory
-jest.mock('#api/Brevo', () => {
-  return {
-    sendContactEmails: jest.fn(),
-  }
-})
 
 describe('ContactFormResolver', () => {
   beforeAll(async () => {
     testServer = await createTestServer()
   })
 
+  beforeEach(async () => {
+    await deleteAll()
+    jest.clearAllMocks()
+  })
+
   describe('createContactForm mutation', () => {
     describe('email is no email', () => {
       it('throws schema error', async () => {
-        const response = await testServer.executeOperation({
-          query: `mutation($data: ContactFormInput!) {
-                    createContactForm(contactFormData: $data) 
+        const response = await testServer.executeOperation(
+          {
+            query: `mutation($data: ContactFormInput!) {
+                    createContactForm(contactFormData: $data)
                   }`,
-          variables: {
-            data: {
-              firstName: 'Peter',
-              lastName: 'Lustig',
-              content: 'Hello DreamMall!',
-              email: 'peter(at)lustig.de',
+            variables: {
+              data: {
+                firstName: 'Peter',
+                lastName: 'Lustig',
+                content: 'Hello DreamMall!',
+                email: 'peter(at)lustig.de',
+              },
             },
           },
-        })
+          { contextValue: mockContextValue() },
+        )
         expect(response.body).toMatchObject({
           kind: 'single',
           singleResult: {
@@ -52,20 +54,23 @@ describe('ContactFormResolver', () => {
 
     describe('email is too long', () => {
       it('throws schema error', async () => {
-        const response = await testServer.executeOperation({
-          query: `mutation($data: ContactFormInput!) {
-                    createContactForm(contactFormData: $data) 
+        const response = await testServer.executeOperation(
+          {
+            query: `mutation($data: ContactFormInput!) {
+                    createContactForm(contactFormData: $data)
                   }`,
-          variables: {
-            data: {
-              firstName: 'Peter',
-              lastName: 'Lustig',
-              content: 'Hello DreamMall!',
-              email:
-                'Aloysius.Bartholomew.Chauncey.Constantine.Devereaux.Ellington.Feliciano.Giuseppe.Horatio.Ignatius@lustig.de',
+            variables: {
+              data: {
+                firstName: 'Peter',
+                lastName: 'Lustig',
+                content: 'Hello DreamMall!',
+                email:
+                  'Aloysius.Bartholomew.Chauncey.Constantine.Devereaux.Ellington.Feliciano.Giuseppe.Horatio.Ignatius@lustig.de',
+              },
             },
           },
-        })
+          { contextValue: mockContextValue() },
+        )
         expect(response.body).toMatchObject({
           kind: 'single',
           singleResult: {
@@ -82,19 +87,22 @@ describe('ContactFormResolver', () => {
 
     describe('first name is too long', () => {
       it('throws schema error', async () => {
-        const response = await testServer.executeOperation({
-          query: `mutation($data: ContactFormInput!) {
-                    createContactForm(contactFormData: $data) 
+        const response = await testServer.executeOperation(
+          {
+            query: `mutation($data: ContactFormInput!) {
+                    createContactForm(contactFormData: $data)
                   }`,
-          variables: {
-            data: {
-              firstName: 'Aloysius Bartholomew Chauncey Constantine Devereaux',
-              lastName: 'Lustig',
-              content: 'Hello DreamMall!',
-              email: 'peter@lustig.de',
+            variables: {
+              data: {
+                firstName: 'Aloysius Bartholomew Chauncey Constantine Devereaux',
+                lastName: 'Lustig',
+                content: 'Hello DreamMall!',
+                email: 'peter@lustig.de',
+              },
             },
           },
-        })
+          { contextValue: mockContextValue() },
+        )
         expect(response.body).toMatchObject({
           kind: 'single',
           singleResult: {
@@ -111,20 +119,23 @@ describe('ContactFormResolver', () => {
 
     describe('last name is too long', () => {
       it('throws schema error', async () => {
-        const response = await testServer.executeOperation({
-          query: `mutation($data: ContactFormInput!) {
-                    createContactForm(contactFormData: $data) 
+        const response = await testServer.executeOperation(
+          {
+            query: `mutation($data: ContactFormInput!) {
+                    createContactForm(contactFormData: $data)
                   }`,
-          variables: {
-            data: {
-              firstName: 'Daenerys',
-              lastName:
-                'Stormborn of the House Targaryen, First of Her Name, the Unburnt, Queen of the Andals and the First Men, Khaleesi of the Great Grass Sea, Breaker of Chains, and Mother of Dragons',
-              content: 'Hello DreamMall!',
-              email: 'daenerys@red-keep.westeros',
+            variables: {
+              data: {
+                firstName: 'Daenerys',
+                lastName:
+                  'Stormborn of the House Targaryen, First of Her Name, the Unburnt, Queen of the Andals and the First Men, Khaleesi of the Great Grass Sea, Breaker of Chains, and Mother of Dragons',
+                content: 'Hello DreamMall!',
+                email: 'daenerys@red-keep.westeros',
+              },
             },
           },
-        })
+          { contextValue: mockContextValue() },
+        )
         expect(response.body).toMatchObject({
           kind: 'single',
           singleResult: {
@@ -141,22 +152,25 @@ describe('ContactFormResolver', () => {
 
     describe('content is too long', () => {
       it('throws schema error', async () => {
-        const response = await testServer.executeOperation({
-          query: `mutation($data: ContactFormInput!) {
-                    createContactForm(contactFormData: $data) 
+        const response = await testServer.executeOperation(
+          {
+            query: `mutation($data: ContactFormInput!) {
+                    createContactForm(contactFormData: $data)
                   }`,
-          variables: {
-            data: {
-              firstName: 'Peter',
-              lastName: 'Lustig',
-              content: `Aus welchen Anlässen hat der Bundesnachrichtendienst (BND) im ersten
+            variables: {
+              data: {
+                firstName: 'Peter',
+                lastName: 'Lustig',
+                content: `Aus welchen Anlässen hat der Bundesnachrichtendienst (BND) im ersten
 Kalendervierteljahr 2022 bei Medien ohne Inanspruchnahme anwaltlicher
 Hilfe um Korrekturen von Berichterstattungen ersuchen lassen (bitte je-
 weils nach Datum, Medium, Anlass und Kosten auflisten)?`,
-              email: 'peter@lustig.de',
+                email: 'peter@lustig.de',
+              },
             },
           },
-        })
+          { contextValue: mockContextValue() },
+        )
         expect(response.body).toMatchObject({
           kind: 'single',
           singleResult: {
@@ -172,20 +186,28 @@ weils nach Datum, Medium, Anlass und Kosten auflisten)?`,
     })
 
     describe('with correct data', () => {
-      it('returns true', async () => {
-        const response = await testServer.executeOperation({
-          query: `mutation($data: ContactFormInput!) {
-                    createContactForm(contactFormData: $data) 
+      const contextValue = mockContextValue()
+      const action = async () => {
+        return testServer.executeOperation(
+          {
+            query: `mutation($data: ContactFormInput!) {
+                    createContactForm(contactFormData: $data)
                   }`,
-          variables: {
-            data: {
-              firstName: 'Peter',
-              lastName: 'Lustig',
-              content: 'Hello DreamMall!',
-              email: 'peter@lustig.de',
+            variables: {
+              data: {
+                firstName: 'Peter',
+                lastName: 'Lustig',
+                content: 'Hello DreamMall!',
+                email: 'peter@lustig.de',
+              },
             },
           },
-        })
+          { contextValue },
+        )
+      }
+
+      it('returns true', async () => {
+        const response = await action()
         expect(response.body).toMatchObject({
           kind: 'single',
           singleResult: {
@@ -196,8 +218,9 @@ weils nach Datum, Medium, Anlass und Kosten auflisten)?`,
         })
       })
 
-      it('calls sendContactFormEmail', () => {
-        expect(sendContactEmails).toHaveBeenCalledWith({
+      it('calls sendContactFormEmail', async () => {
+        await action()
+        expect(contextValue.brevo.sendContactEmails).toHaveBeenCalledWith({
           firstName: 'Peter',
           lastName: 'Lustig',
           content: 'Hello DreamMall!',
@@ -206,6 +229,7 @@ weils nach Datum, Medium, Anlass und Kosten auflisten)?`,
       })
 
       it('writes event to database', async () => {
+        await action()
         const result = await prisma.event.findMany()
         expect(result).toHaveLength(1)
         expect(result).toEqual([
