+++ backend/src/use-cases/leave-table.ts
@@ -1,6 +1,6 @@
 import { deleteTable } from './delete-table'
 
-import type Logger from '#src/logger'
+import type { Logger } from '#src/logger'
 import type { PrismaClient } from '#src/prisma'
 
 const oneElseIsModerator = (userId: number) => (user: { userId: number; role: string }) => {
@@ -8,7 +8,7 @@ const oneElseIsModerator = (userId: number) => (user: { userId: number; role: st
 }
 
 export const leaveTable =
-  ({ prisma, logger }: { prisma: PrismaClient; logger: typeof Logger }) =>
+  ({ prisma, logger }: { prisma: PrismaClient; logger: Logger }) =>
   async ({ userId, tableId }: { userId: number; tableId: number }) => {
     const meeting = await prisma.meeting.findFirst({
       where: {
