+++ backend/test/mockContextValue.ts
@@ -1,30 +1,21 @@
 import { prisma } from '#src/prisma'
 
+import { createMockConfig } from './mockConfig'
+
+import type { createBrevoClient } from '#src/api/Brevo'
 import type { Context } from '#src/context'
 
 export const mockContextValue: (overrides?: Partial<Context>) => Context = (overrides = {}) => {
+  const config = createMockConfig()
+  const brevo = {
+    sendContactEmails: jest.fn(),
+    subscribeToNewsletter: jest.fn(),
+    confirmNewsletter: jest.fn(),
+  } as ReturnType<typeof createBrevoClient>
   const defaults: Context = {
     user: null,
-    config: {
-      BREVO_KEY: '',
-      BREVO_ADMIN_NAME: '',
-      BREVO_ADMIN_EMAIL: '',
-      BREVO_CONTACT_TEMPLATE_ADMIN: 1,
-      BREVO_CONTACT_TEMPLATE_USER: 2,
-      BREVO_NEWSLETTER_TEMPLATE_OPTIN: 3,
-      BREVO_NEWSLETTER_LIST: 3,
-      BBB_SHARED_SECRET: '',
-      BBB_URL: '',
-      BBB_PULL_MEETINGS: '',
-      BBB_WEBHOOK_URL: '',
-      FRONTEND_URL: '',
-      JWKS_URI: '',
-      WELCOME_TABLE_NAME: '',
-      WELCOME_TABLE_MEETING_ID: '',
-      SENTRY_DSN: '',
-      SENTRY_ENVIRONMENT: '',
-      WEBHOOK_SECRET: undefined,
-    },
+    config,
+    brevo,
     dataSources: { prisma },
   }
   return {
